Work on dht11 integration

+++ Ranger.hpp
@@ -4,6 +4,6 @@ float avgRange(int, bool debug = false);
 float medianRange(int, bool debug = false);
 void calibrate();
 
-extern float humidity;
-extern float temp;
+extern int humidity;
+extern int temp;
 extern float c;
